+++ sdk/network/azure-mgmt-network/generated_tests/test_network_management_bgp_service_communities_operations.py
@@ -22,7 +22,7 @@ class TestNetworkManagementBgpServiceCommunitiesOperations(AzureMgmtRecordedTest
     @recorded_by_proxy
     def test_bgp_service_communities_list(self, resource_group):
         response = self.client.bgp_service_communities.list(
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
         result = [r for r in response]
         # please add some check logic here by yourself