+++ sdk/network/azure-mgmt-network/generated_tests/test_network_management_virtual_appliance_skus_operations.py
@@ -22,7 +22,7 @@ class TestNetworkManagementVirtualApplianceSkusOperations(AzureMgmtRecordedTestC
     @recorded_by_proxy
     def test_virtual_appliance_skus_list(self, resource_group):
         response = self.client.virtual_appliance_skus.list(
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
         result = [r for r in response]
         # please add some check logic here by yourself
@@ -33,7 +33,7 @@ class TestNetworkManagementVirtualApplianceSkusOperations(AzureMgmtRecordedTestC
     def test_virtual_appliance_skus_get(self, resource_group):
         response = self.client.virtual_appliance_skus.get(
             sku_name="str",
-            api_version="2024-05-01",
+            api_version="2024-07-01",
         )
 
         # please add some check logic here by yourself